cinema/pyproject.toml
[tool.poetry]
name = "cinema-drf"
version = "0.1.0"
description = ""
authors = ["olchrkk <olchrkk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.1"
pillow = "^10.2.0"
psycopg-binary = "^3.1.17"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.9"
django-ckeditor = "^6.7.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
